[fix] Appropriately modify `renderChunksMany` when toggling Freecam
To optimise chunk rendering, minecraft tries to omit (normally) invisible chunks from chunk rendering. However, when the player's head is in an opaque block AND the classic 'block in head' overlay is not drawn, the player should see a sort of x-ray like view of the world. This happens in e.g. spectator mode. The `renderChunksMany` field in Minecraft also controls this behaviour, allowing modders to turn off the optimisation.