[fix] F3 Debug Pie would crash on Single Player
Possibly would happen on other instances, we will want to test this some before pushing to master/release.

EntityPlayer was sending PlayerTravelEvent, this caused the EventProfiler handler to concurrently modify the Minecraft.profiler instead of the MinecraftServer.profiler.

Additionally, this caused problems when trying to test anything using PlayerTravelEvent as it was getting more than what it should have.